Clean up Vagrantfile remove .vagrant dynamic files
Added some common stuff to the .gitignore so that git would not
track generated or compiled files, those don't belong in source
control.
Modified the Vagrantfile to install lxml in the virtualenv
using the requirements.txt as opposed to using the system
package. This decouples your app from your platform, which is
important if you ever wanted to deploy on another one, like GAE
(a good candidate for this project BTW).
Also organized the repo a little better.